requirements.txt
# ─── Core Processing ───────────────────────────────────────────────────────────
pyspark==3.5.1

# ─── Orchestration ─────────────────────────────────────────────────────────────
apache-airflow==2.9.1
apache-airflow-providers-apache-spark==4.8.0

# ─── Data Generation & Manipulation ────────────────────────────────────────────
faker==24.11.0
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

# ─── Data Quality ──────────────────────────────────────────────────────────────
# (built-in DQ framework — no great-expectations needed)

# ─── Config & Utilities ────────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1

# ─── Audit Logging ─────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ─── Query layer (view Parquet tables locally without Spark) ───────────────────
duckdb==0.10.2

# ─── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-cov==5.0.0
chispa==0.9.4
